FEATURE Matomo 4 support
