+++ plugin.json
@@ -1,16 +1,17 @@
 {
   "name": "LdapVisitorInfo",
-  "version": "3.0.1",
+  "version": "4.0.0",
   "description": "Piwik plugin to get visitor details (thumbnail, description) from LDAP",
   "keywords": ["Intranet", "Visitor", "LDAP", "ThaDafinser", "thumbnail"],
   "license": "GPL v3",
   "homepage": "https://github.com/ThaDafinser/Piwik-LdapVisitorInfo",
   "authors": [
-    {"name": "Martin Keckeis", "homepage": "https://github.com/thadafinser"}
+    {"name": "Martin Keckeis", "homepage": "https://github.com/thadafinser"},
+    {"name": "Dmitrii Shiriaev", "homepage": "https://github.com/dshiryaev-plesk"}
   ],
   "theme": false,
   "require": {
-    "piwik": ">=3.0.0-b1,<4.0.0-b1",
+    "piwik": ">=4.0.0-b1,<5.0.0-b1",
     "php": ">=7.0.0,<8.0.0"
   }
 }
